+++ src/pages/Main.jsx
@@ -1,10 +1,8 @@
 import Layout from '../components/Layout';
-import Header from '../components/Header_';
 
 const MainPage = () => {
   return (
     <>
-      <Header />
       <Layout>
         <h1>MainPage</h1>
       </Layout>
