+++ src/router/MainRouter.js
@@ -5,10 +5,12 @@ import PostPage from '../pages/Post';
 import CreatePostPage from '../pages/CreatePost';
 import Login from '../pages/Login';
 import Signup from '../pages/Signup';
+import Header from '../components/Header';
 
 const MainRouter = () => {
   return (
     <BrowserRouter>
+      <Header />
       <Routes>
         <Route path="/" element={<MainPage />} />
         <Route path="/post">
